Move Merge options to new admin submenu
